+++ backend/app/app_models/db_models.py
@@ -1,12 +1,8 @@
 from datetime import datetime
-
-from pydantic import BaseModel
 from sqlalchemy import Column, Integer, Float, DateTime
 from sqlalchemy.ext.declarative import declarative_base
 
-
 Base = declarative_base()
-
 
 class SensorData(Base):
     __tablename__ = "sensor_data"
@@ -16,10 +12,3 @@ class SensorData(Base):
     temperature = Column(Float, nullable=False)
     pressure = Column(Float, nullable=False)
     timestamp = Column(DateTime, default=datetime.utcnow)
-
-
-class SensorDataIn(BaseModel):
-    sensor_id: int
-    timestamp: float
-    temperature: float
-    pressure: float
