+++ backend/app/app_models/schemas.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class SensorDataIn(BaseModel):
+    sensor_id: int
+    timestamp: float
+    temperature: float
+    pressure: float
