+++ backend/app/db_layer/session.py
@@ -1,0 +1,8 @@
+from db_layer.database import SessionLocal
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
